+++ app/Transformer/Api/AccountTransformer.php
@@ -34,7 +34,8 @@ class AccountTransformer extends Fractal\TransformerAbstract
 			'local' => (bool) $local,
 			'is_admin' => (bool) $is_admin,
 			'created_at' => $profile->created_at->toJSON(),
-			'header_bg' => $profile->header_bg
+			'header_bg' => $profile->header_bg,
+			'last_fetched_at' => optional($profile->last_fetched_at)->toJSON()
 		];
 	}