+++ resources/lang/th/timeline.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+
+  'emptyPersonalTimeline' => 'ทามไลน์ของคุณยังไม่มีอะไร!',
+
+];